feat: finished forgot password #84
