+++ ASM2_API/ASM-Java5/src/main/java/com/dto/auth/EmailRequest.java
@@ -1,0 +1,11 @@
+package com.dto.auth;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+
+public record EmailRequest (
+        @NotBlank(message = "Email is required")
+        @Email(message = "Please provide a valid email address")
+        String email
+){
+}
