+++ ASM2_API/ASM-Java5/src/main/java/com/dto/auth/OtpRequest.java
@@ -1,0 +1,9 @@
+package com.dto.auth;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record OtpRequest (
+        @NotBlank(message = "OTP is required")
+        String otp
+){
+}
