+++ ASM2_API/ASM-Java5/src/main/java/com/util/TokenGeneratorUtil.java
@@ -15,4 +15,7 @@ public class TokenGeneratorUtil {
         secureRandom.nextBytes(randomBytes);
         return base64Encoder.encodeToString(randomBytes);
     }
+    public String generateOtp() {
+        return String.format("%06d", secureRandom.nextInt(1_000_000));
+    }
 }
